requirements.txt
Flask
python-dotenv
python-docx
Markdown
openai 